Stripe Activity to reuse UI code (#337)
* small example

* address commentss
