+++ example/res/layout/activity_launcher.xml
@@ -69,13 +69,4 @@
         android:text="@string/addAddress"
         />
 
-    <Button
-        android:id="@+id/btn_customer_session_launch"
-        android:layout_width="200dp"
-        android:layout_height="wrap_content"
-        android:layout_margin="10dp"
-        android:layout_gravity="center"
-        android:text="@string/launch_customer_session"
-        />
-
 </LinearLayout>
