+++ stripe/AndroidManifest.xml
@@ -2,18 +2,29 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           package="com.stripe.android"
           android:versionCode="16"
-          android:versionName="5.0.0">
+          android:versionName="5.0.0"
+    >
 
     <application>
         <activity
             android:name=".view.AddSourceActivity"
             android:theme="@style/StripeDefaultTheme"
-            android:windowSoftInputMode="stateVisible">
+            android:windowSoftInputMode="stateVisible"
+            >
+
         </activity>
 
         <activity
             android:name=".view.PaymentMethodsActivity"
-            android:theme="@style/StripeDefaultTheme">
+            android:theme="@style/StripeDefaultTheme"
+            >
+            >
+        </activity>
+
+        <activity
+            android:name=".view.AddAddressActivity"
+            android:theme="@style/StripeDefaultTheme"
+            >
         </activity>
     </application>
